Add full-text search to `Events` view
This allows the user to search for words in the `title` and `description` of an event. To achieve this, I have added an exposed filter in the view and updated the search API to index titles and descriptions. I have also unenabled case sensitivity.

To get the right design, I have added some template suggestions in the `norvel.theme` file and changed the order of filters in the view template.