Add full-text search to `Articles` view
This allows the user to search for words in the `title` and `description` and `teaser` of an articles. To achieve this, I have added an exposed filter in the view and updated the search API to index titles and descriptions and teaser text. I have also unenabled case sensitivity.

And aligned the advance settinges for the view to user better expose fileters with the settings from events view