Fix for generated files
Some static files needs to be generated by Drupal the first time they
are requested. This relies on the try_files directive which redirects
requests for non-existent files to Drupal.

It turns out that any headers added to the response beforehand are
lost in this process. To address this issue we add a new location
which can be used in such situations. This new location is a copy
of the existing @drupal location with the added header.

We could in theory also have added it with an if-statement for the
request path in the existing location but that would spread the
handling across multiple configuration files. This approach has some
duplication but all handling is colocated within a single file.