Implement user message on deletion of an opening hours category
Now, when a user deletes a category, a message will display indicating the number of associated opening hours.

To facilitate this, the `$categoryId` parameter was added to the `loadMultiple` method in the `OpeningHoursRepository`.

At present, no specific action occurs on term deletion. Handling this will be part of the next commit.