Optimizing GH actions, to use less minutes. (WIP)
We can use `.lagoon.yml` to set up a GH deployment,
that we can use to listen for in GH actions.
This is an alternative to us spending a lot of GH minutes
just waiting for the site to become available.

As the `@todo`'s say, this needs some work, as I'm unsure
how to use secrets and get around `set -e` as part of
the deployment.
**I'm looking for input for this :)**

Ontop of that, also setting up `concurrency` rules to
`ci-tests`, so if a second push is made, we cancel the
old and unrelated workflow.