Fix url to React components release 2024.39.0
The file currently mentioned does not exist. The appended SHA differs
from what is attached to the release.

Why this error occurred is unknown.

The current changes are the result of running
RELEASE=2024.39.0 VERSION=2024.39.0 task dev:composer:update-react

Note that this also rolls back a lot of changes made to the lock file
when design system and React components where added in
6f7a3f40c1f4f5f5951fb3c2dccdf5b07d4fbc45