Revert event reoccurrence handling
This reverts commit 20ac720dca95d86c268ec30d5079be82cb9e8c40, reversing
changes made to 51abf7e8f24515bb14653a5bafc4605f6b3267b7.

When rolling out these changes to existing sites this causes problems
during the configuration import part when running drush deploy:

> Error: Call to a member function getLabel() on null in /app/web/modules/contrib/scheduler/src/SchedulerManager.php on line 1171 #0 /app/web/modules/contrib/scheduler/scheduler.module(1450): Drupal\scheduler\SchedulerManager->entityUpdate()
> #1 /app/web/core/lib/Drupal/Core/Config/ConfigImporter.php(574): _scheduler_config_import_process_entity_definitions()
> #2 /app/vendor/drush/drush/src/Commands/config/ConfigImportCommands.php(259): Drupal\Core\Config\ConfigImporter->doSyncStep()

git bisect has traced the error back to this change. Reverting it
removes the problem.

Based on this we cannot tell if it is the update of the module
which was introduced as a part of the updated event reoccurrence
handling which causes the problem or the actual event reoccurrence
handling.

# Conflicts:
#	composer.lock