Display branch address on series level. DDFHER-38
We have a custom field formatter for the address field, that pulls the
address of the branch if no alternative is set.
However, the code only worked for eventinstances, and not eventseries.
I've fixed it, so now, it happens for both.
This also meant moving some code from the `EventWrapper`, as that also
only works with EventInstances.
..And that meant that I had to update the Rest API, to use the field
formatter, rather than pulling data directly