Allow retrieval of opening hours for multiple branches (redia_legacy)
Updated the `OpeningHoursRepository::loadMultiple` method to accept  an array of branch IDs, enabling queries for multiple branches using the `IN` operator.

In OpeningHoursResource, the `branch_id` / `nid` parameter now accepts a single ID or a comma-separated list of IDs.

These changes maintain backward compatibility while enhancing the `redia_legacy` API to support retrieval of opening hours for multiple branches in a single request.