Also create deployments when PRs are synchronized
Currently we only create deployments when pull requests are created
or reopened. As a result the environment will be shown as outdated
if there are additional commits pushed to the branch.

Update the workflow to also run on synchronisation to address this.

This was the same approach as used previously.