Add support for raw formatting ranges of prices
Sometimes we want to display ranges of prices in situations where it
is not appropriate to apply the additional formatting that we normally
do (current prefix, "free" instead of 0 etc.)

One example for this is the RSS feed for the Redia app.

This adds support for this supported by a unit test.