Temporarily disable the RSS feed for exposing events to Redia
Problems with the current feed has led to events appearing in
problematic ways e.g. by being listed as free to attend but actually
having non-free ticket categories registered.

To avoid this we temporarily disable the feed to avoid exposing data.

Provide a cacheable response with a HTTP error code stating that the
feed is not available currently.

404 seems like an appropriate error code given the situation even
though this is strictly not a client issue. 503 indicates a problem
which is more temporary than is the case there. 501 indicates problems
with the HTTP method which is not the case.

To make static analytics tools happy we have to comment out and adjust
a bit of the surrounding code as well.