Remove `Safe\array_combine` and use native `array_combine` instead
In PHP 8.0 and later, the native array_combine function has been enhanced to throw a ValueError exception when the number of elements in the keys and values arrays do not match, aligning its behavior with the Safe\array_combine function. Consequently, the Safe\array_combine function is deprecated in these versions.
PHP