Update `drupal/focal_point` to "2.1.1"
The previous patch `3162210` has been integrated into `drupal/focal_point` version "2.1.1" and has been removed from the project.

Following the update, a new issue was identified where the preview window fails to open. A patch `3462165` has been created to resolve this issue.

Refer to the video in the PR for additional context.