Added a service decorator for simple_oauths DisallowSimpleOauthRequest service.
In our decorator, we added functionality that checks wether the request path
starts with '/graphql' or not.

The reason for this is, that right now the isOauth2Request function only checks
if the request contains a bearer token. This is also the case when the requests
come from the opening hours REST API endpoints, as it submits a library token.

By checking for the /graphql in the path, we make sure to only apply the oauth2
validation on the graphql requests.