Updated dpl_consumers module so it prefills the new consumer fields that the simple_oauth module provides upon installation.
Right now we hardcode the secret to "secret", but we will have to find
a better way of setting this value on installation. Maybe through an
environment variable in Lagoon?

Created a new drush command for setting the GraphQL consumer secret,
based on what is set in .env.

Also created a new task in the Taskfile.yml, used for running our new
drush commando for setting a secret locally.