Fixing `develop` issues, due to drush being locked in `dpl_graphql`
Copying the `dpl_graphql` code to a seperate, new `dpl_graphql2` module.
This is necessary, as the original graphql module includes code that does
not work without simple_oauth, which cannot be installed,
due to the Drush command not being able to run.
We get around this, by moving the code into a new module (graphql2),
leaving the old module as an empty boilerplate.
This gives drush a chance to actually enable the dependencies.
In a future release, we should rename the module again.