Patch Drupal Core to prevent errors when decorating non-existent services
We currently have an issue with our DPL Graph module referring to
a service provided by the simple_oauth module. We would normally
enable this module using an update hook (or config import) but it is
not possible to do so as Drush will error out befure runninng
anything.

To address this we patch Drupal Core to add support for the Symfony
decoration_on_invalid property.

This patch has been accepted for Drupal 10.4.x and 11.x so it should
be safe to merge.