Create BNF UUID field on articles.
This field will be used to keep track of which articles have been
imported to/from BNF already.
The importer will take the original node UUID and place it here.
Technically, this will be a read-only field, as we never want the user
to actually edit it, but I still decided to introduce it as a regular
node field as it will make debugging in the future much easier if we
have access to it, through the Drupal interface.