BNF Server: Import Request graphQL endpoint.
Creating a custom mutation graphQL endpoint, as part of the `bnf_server`
module.
This endpoint will receive a ping from client libraries, when they want
a certain node of theirs to be imported to BNF.
The client will send along a UUID and an endpoint which BNF can use to
do a GraphQL query against, to get the node data.