Editor interface for using BNF Client export.
Adding a simple form field to the node edit form, with a checkbox.
If the checkbox is checked, the node will be published and sent to BNF
upon save.
Originally, I wanted to have a "publish to BNF" button, but that became
very complicated, as I'd have to replicate some of the submit handling
that happens with the core "submit" button.
In the end, I think this solution is user friendly enough.