Display 'import to my site' button on BNF server. DDFHER-181
If the editor has been redirected to the server from the BNF login, the
context from the site they come from, has been saved to a cookie.
If that is the case, we show a floating (fixed) button on nodes, that
links back to the original site, along with the UUID.
If the article has already been imported, the original site will display
a warning.