Increase size of NGINX persistent storage to 10Gi
This mirrors a change introduced for the site templates in
dpl-platform.

While most core sites will not need the additional storage increasing
the storage limit will not introduce additional costs. Having the
same limit as production sites ensures that we avoid hitting the limit
if we need to synchronise data from a production site to a PR site.