Use latest version of Go
We cannot see the direct cause of why the node container fails to build.
It is often because of mismatch between what is needed of environment
variables and what is available.
Using the latest version of Go probably solves the problem.
Otherwise we could comment the Go application out if it keeps being
obstructive.