Refactor `go_linkbox` paragraph to support a single link
I encountered issues when trying to remove the `field_link` from the `go_link` paragraph. As a result, I deleted it and recreated it all. The main issue stemmed from a previous reuse of the `field_link`, which allowed multiple links. To resolve this, I have introduced a `field_go_link` that supports only one link.