Patching recurring_events to automatically clean up orphaned events. DDFHER-204
If for whatever reason an eventinstance ends up orphaned after a series
has been deleted, the event API and the breadcrumb builder will fail
hard.

I've created an upstream patch at
https://www.drupal.org/project/recurring_events/issues/3504806 that
deals with this - it cleans up orphaned eventinstances as part of cron.

The custom code problems will be fixed in a seperate commit.