Make custom event code not fail with orphaned events. DDFHER-204
recurring_events has a ->getEventSeries() function that always should
return a series, but sometimes it doesnt.
As part of a previous commit, we make sure that orphaned eventinstances
get removed, but until they are removed, the breadcrumb builder and
event API will fail.
This commit makes our code more robust when it stumbles upon an orphaned
instance.