Remove logic that disallows editing standalone eventinstances. DDFHER-194
We have previously made some code, that redirects the editor when trying
to edit an eventinstance that has no other siblings.
The idea was that we'd prefer the editor to add details on the series
level, so the data can be re-used.

However, this is a problem when wanting to edit the dates.
If the dates are edited directly on the series, it will delete all the
instances and recreate them, destroying any custom changes and causing a
new ID to be used in the event API.
Instead, we'll remove this logic, and depend on the notice that is
already displayed when editing the instance.

Ontop of that, we add a patch to `recurring_events` that adds a notice
link when trying to edit the dates of a series level.
This patch is somewhat ugly, but it is the only way we can add the
notice, as the contrib code is heavily nested, and cannot be altered by
a custom hook.