Remove `field_link` and add `field_go_link_info`
Due to a deployment issue, we had to revert https://github.com/danskernesdigitalebibliotek/dpl-cms/pull/2030 This restores the functionality.

There is still an issue where we cannot rename or replace (delete and add) the `field_go_link` field in the `go_linkbox` paragraph.

To address this, the `field_link` field, which allowed multiple links, has been removed and replaced with a new field, `field_go_link_info`, which only supports a single link.

I’m not entirely satisfied with the name and would have preferred calling it `field_go_link`. But that was not possible.