Add and enable HTTP cache control module
To control caching in the browser we need to distinguish between
caching in Varnish and caching in the browser. We want the cache
lifetime in Varnish to be long because we explicitly purge entries
with the Purge module based on content changes. We want the cache in
the browser to be short(er). This is a compromise between showing
up to date content and ensuring good perceived performance by reducing
client/server communication.

Currently the cache time for both is controlled through the maxage
directive in the Cache-Control HTTP header. This is configured by
Drupal Core. To separate we want to add support for the separate
s-maxage directive which is targeted shared caches such as Varnish.

We can do this by adding the HTTP Cache Control module. This provides
more complex control over the Cache-Control header while being
consistent with the Core implementation. This means we do not manage
the Cache Control header multiple places.

The module seems well maintained. It is widely used, with few open
issues, has familiar backers, is updated and supports Drupal 11.

Consequently the module is added and enabled with the default
configuration.