Move cache max age to shared max age
We want reduce caching in the browser but preserve caching in Varnish.

To do this we move the current cache max age of 1 month to the
shared max age (s-maxage) and reconfigure the max age to 5 minutes.

There is no project specific reason behind the value of 5 minutes. It
is the recommended value in the Drupal performance UI. This seems to
be somewhere between always shown updated content from editors and
ensuring a good perceived performance by end users.