Set Canonical URL, as part of BNF sync. DDFHER-187
When exporting to, or importing from BNF, we want the canonical URL to
be sent along.
First of all, we will respect if any data already exists in the field -
but if nothing exists, we add the origin URL of the node.
Creating the origin URL is a bit annoying - we need to get the base URL
of the site, as part of `info`, and then put it together with the `path`
of the node.
If the original node has its path updated, `pathauto` automatically
creates a redirect.