Move configuration handling away from library token handler
Since the handler now accepts agency id, client credentials and token
endpoint as arguments there is no need to also pass this as a
constructor argument. It is only used sometimes.

This allows us to simplify the library token handler.

The logic for using the configured values is moved to the cron
implementation as this is the primary situation we want to use the
values stored in the configuration.

This change also allows us to simplify the related unit test:
- Setting up the token handler requires less prophesizing
- Test cases dealing with configuration are no longer relevant