Add Drupal as a dependency for Mapp script
Our Mapp integration script uses Drupal.behaviors. If this is not
defined then errors will be thrown. For whatever reason this has not
occurred previously but it does now.

Deactivating JS aggregation fixes this but it is not a long term
solution at the moment.

To address this we add the core/drupal library as a dependency.