Add helper function to unignore configuration entries
Sometimes we modify configuration on sites runtime as a temporary
measure to address a problem. As we us the config_ignore_auto module
this configuration will now be ignored from future configuration
imports.

In many cases we will want to update the standard configuration to
address this problem. Consequently we need to unignore these
configuration entries again. This is likely to be done in an update
hook.

This adds a helper function which makes it easier to remove an
ignored entry from the config ignore auto module.