Ensure generated GraphQL client has not drifted from VCS
Regenerate code from GraphQL queries and check whether this results
in any deviations from what is currency committed.

This helps identify changes to our GraphQL queries that developers may
have forgotten to commit or changes to third party code that affects
the generated code.

Both situations should be raised related to the change instead of the
next time around when a developer generates client code in relation
to other work.

This approach is consistent with what we do for other generated code.