Re-enable and re-export all GraphQL fields with Sailor
This enables all fields and paragraphs as part of the BNF-supported
content types (article, page, go_article, go_page, go_category), and
re-exports it using sailor.
Some of the fields have changed (such as name and `required`), so some
of the mappers needed to be updated.
I've also updated the `::make`'s in the testers to have named arguments,
as we cannot assume the order in `node.graphql`, as it makes it
impossible to manage.
DDFHER-168