BNF: Skip unsupported paragraphs when mapping.
Instead of failing the whole node import, we want to skip past the
paragraphs that do not have valid mappers.
In the long run, we also want to display some kind of notice for the end
user, but for now, we're adding it as a log in the BNF site.