Set sorting by `changed` on `/search/web/TERM` page.
The category/tag search, that is shown on the canonical page, is
currently being sorted by `relevance`, just like the  `/search/web` page
- but, it makes no sense, as there is no user input.
Instead, we'd like this page to be sorted by the latest updated content
(`changed`).
This includes creating an aggregrated field in search API, as we both
have eventseries and node entity types.
DDFSAL-71