Remove diff check for GraphQL schema changes
For whatever reason the schema generated in CI has elements in a
different order than what is exported locally. Since this difference
does not have any meaning regarding the sematics of the schema we
disable the check for now.

For now the check that the generated code remains is sufficient.

Later on we may look into a more intelligent approach for detecting
changes to the schema.