Updated the Paragraph field on the Article, Branch and Page content type.
Instead of specifying an exclude list of paragraphs that should not be available,
we now instead use an include list to specify which paragraphs should be available.

By doing this we make sure that all content types that include a Paragraph field, are
now using the include list to specify which paragraphs is available.