Change format of library token expire attribute
And make sure that both library and user token are uncached.

We need to make sure that the tokens are always fresh.
Caching could be improved by using context, but we go with uncached for
now.

The expire format of the user token has changed to a DateTime object, so
it has the same signature as the library token. Also we make sure that
every time we have an expire time it is the exact date where the token
is expiring.