Implement a cache of node types in the path processor
The consequence of path rewriting is that we have to load each linked
node to check its type to determine whether we need to rewrite the
URL.

In order to limit the number of entity loads we do, we implement a
static cache to keep track of whether a nid is a go node or not. We
save this cache to state (as even CACHE_PERMANENT cache items are
flushed on cache rebuild) and do it in chunks of a hundred items to
keep down the amount of queries when our cache is filled.