Enabled the "View preview entities" permission for the GO GraphQL Client role.
This makes it possible for our "go_graphql" user to use the preview functionality
in graphql.

For this to work, we have to use a "preview_token". The redirect token is now set
as part of the redirect response. The GO FE can then fetch the preview token from
the URL, and make the correct GraphQL query.