Fix getting images from DT to CMS
When doing GraphQL requests from CMS to DT, Drupal used the wrong site
settings file due to `bnfnginx` not matching `bnf-`. Didn't have an
influence on the returned data as it still used the right database
service, but it had the effect that it thought that files was in
`site/default/files`.

Fix the `sites.php` hackery to account for this.