Create new Material Grid paragraph base. DDFSAL-161
This creates a new paragraph named `material_grid_automatic_v3`, that
will be used as a replacement for both `material_grid_automatic` and
`material_grid_link_automatic`.
In an attempt to make it easier to review, this commit just adds the
base paragraph, without the actual material fields - e.g. nothing we
haven't seen before.