Create material_search fieldWidget for material grid paragraph. DDFSAL-161
This introduces a `field_material_search` field, on the new
`material_grid_automatic_v3` paragraph.

This field comes with a new, custom FieldWidget and FieldType.
It has value fields for the "link-to-a-search-page", CQL, and various
material search filters.
The custom FieldWidget has some AJAX logic, that fills out the other
fields, when the user inserts a valid search link.