Disable advanced CQL Search fields on GO paragraphs. DDFSAL-161
The `CqlSearch` field type includes some new fields, such as `location`.
These fields are only supported by `material_grid_automatic`, but the
fieldtype is also in use with `go_material_slider_automatic` and
`go_video_bundle`.
To avoid confusion for the editor, we'll hide the advanced fields for
the GO paragraphs, done by a custom FieldWidgetSetting.