Rework cache clearing when installing sites
For whatever reason drush cache:rebuild fails with the error
"The node entity type does not exist." when running drush ci:reset
locally.

The situation seems to fix itself when running drush deploy. drush
deploy internally also rebuilds the cache so there is little reason
to do this explicitly.

Instead we only clear the external cache which requires a separate
command. We do this after drush deploy for good measure.