Find duplicate medias, and hide them in the media library. DDFSAL-155
As part of a bug, all medias pulled from Delingstjenesten/BNF were
duplicated resulting in a bloated media library.

This includes a migration action, where we unpublish all medias that
have been indicated as both duplicated and orphaned.

We choose *not* to delete, to be on the safe side. We mark the medias
with a revision log message, that we can use to look it up and delete
them in a future deploy.