Do *NOT* set BNF-imported nodes as entity templates. DDFSAL-174
When a node is BNF-imported, we do *NOT* want it to show up as a node
template on the /node/add page. Apparantly, the entity_clone_template
module assumes that no value must mean TRUE.

This commit also includes cleaning up the existing nodes that already
have been imported, and has been set as templates.
The client will be informed of this update.