Improve error handling when importing nodes
Currently we do not check for GraphQL errors in the response when
importing a node.

This leads to conditional code and generic error messages which can
be hard to debug.

Add the errorFree() check to the result data retrieval to ensure that
an exception is thrown if the result contains errors. This will bubble
to our general exception handling for the importer which includes the
exception message.