Support configuring a first accession date filter for CQL search field
Filtering by first accession date consists of a date and an operator.

The date can be an actual date in the format YYYY-MM-DD or a relative
date in the format NOW - DAYS/MONTHS.

The operator currently supports >,= and < as provided by the enum.

Add these values to the schema for the field, as properties for the
field type and as columns in the database schema.

Add an update hook to add the columns to existing installations.