Add view for managing content using firstaccessiondate in CQL searches
Support for using firstaccessiondate in CQL searches will be removed.
Getting an overview of where this feature is used can be complicated.

To help editors identify content which needs to be updated we add a
view which shows content related to Material Grid Automatic paragraphs
using the CQL search field where the CQL search string contains
firstaccessiondate.

This is a temporary solution to the problem so this view is not linked
into the standard administration interface. If such features prove
useful in the feature we will discuss how to integrate them into the
existing structure.