Add GitHub action that checks if update hooks has been used. DDFSAL-156
As we allow webmasters to add their own modules, and alter their own permissions, we cannot just update these through config files.
This action checks if any changes has been made in the config files for modules (core.extension.yml), or for permissions (user.role.*.yml) and then checks if the correct `dpl_update` hooks has been added.

This is not a bulletproof method, but it is also only considered a backup solution for something that should be caught in a review anyway.