Breadcrumb and URL alias should respect local time.
If an editor has set an event to start at 00:00, the value saved in the
database is in UTC time - e.g. 22:00 the day before.
The breadcrumbhelper creates the breadcrumb, and the URL alias, but due
to a bug, it generates the string based on UTC.

This has not been noticed before, as we do not display the time as part
of the string, but it is noticable when UTC actually is the day before
as the 00:00 example.