Update dpl_po import command to correctly clear caches. DDFSAL-86
We have a command that is run on all sites regularly, looking something
like this:

```
drush dpl_po:import-remote-config-po da
https://danskernesdigitalebibliotek.github.io/dpl-cms/translations/da.config.po
```

However, we are experiencing that the field labels are not always
translated - despite the translation actually being loaded into Drupal.

After a lot of debugging, it has been clear that the relevant caches
need to be cleared as part of the command being run - otherwise, the
translations may not apply, even with a future cache clear.