'All day' events should be respect in API. DDFSAL-255
We previously used a `recurring_events` hook to alter 'all day' events to be 00:00 - 23:59.
Unfourtantely, this does not work when all you do is enable 'all day' on a series: The hook is not run.

Instead, we do the logic on `eventinstance_presave()`, which is triggered, even when non-date info is changed on the series level.

Another bonus about this, is that it is also respected when editing an instance directly.