Fix branch creation during PR creation
Currently our workflow to create PRs with updates to the design system
fails with the following error:

fatal: invalid reference: advanced-search-v2-fix-suggest
Error: Invalid status code: 128
    at ChildProcess.<anonymous> (/home/runner/work/_actions/stefanzweifel/git-auto-commit-action/v7/index.js:17:19)
    at ChildProcess.emit (node:events:508:28)
    at maybeClose (node:internal/child_process:1101:16)
    at ChildProcess._handle.onexit (node:internal/child_process:305:5) {
  code: 128

This seems to be a result of upgrading
stefanzweifel/git-auto-commit-action from v5 to v7.

Create the branch by explicitly setting create_branch to true. This
may the result of a new default value.