Clear the GraphQL AST cache at the end of deployments
We occasionally see the GraphQL module in combination with GraphQL
Compose not returning a schema with all the custom modifications from
the DPL GraphQL module.

Practice has shown that the situation can be resolved by running
drush cc bin graphql.ast.

This is likely the result of some kind of race condition where the
schema is generated and cached before our custom module has had a
chance to modify it. Clearing the cache manually is a simple way to
get around this.

Insert the clearing here to mirror what we do for the production sites
in
https://github.com/danskernesdigitalebibliotek/dpl-platform/blob/a8b406d4a4c382e4b79829b0b7f720c0919dbe1e/infrastructure/dpladm/env-repo-template/standard/.lagoon.yml#L64-L64

The issue is discussed in
https://www.drupal.org/project/graphql/issues/3477239.