Move more references from dplplat01 to dplplat02
Currently we are still seeing references in GitHub to dplplat01 for
deployents.

This may be the result of our scripting in Lagoon YAML. Update it
and a few other references from dplplat01 to dplplat02.

This intentionally does not include references to Go domains the
include-listing for UniLogin urls is still tied to dplplat01.