Create event_location_type field, allowing marking online events. DDFSAL-449
As part of an admin-interface redesign of the events, we include
location type field, where the editor can choose if an event is physical
or online.

Certain fields (such as address) will be hidden when choosing online,
and the address formatter will also take over in that case.

The field is NOT required on eventinstances, as the N/A option allows us
to correctly use field inheritance. This is something we might want to
introduce to other checkbox fields in the future.